Abbreviations made easy!
nomencl package added to make it possible to use the command \nomenclature{Abbr}{Discription} in the text to automatically add abbreviations and a line or two describing them to the Abbreviations-list in the report.

To build the list MakeIndex must be run with the line "%bm.nlo -s nomencl.ist -o %bm.nls", where %bm is the file name without file extention.

Compiling without building the abbreviations list is still possible without
problems.